refactor: extract out App model
